// console/types/actions.ts
export enum Actions {
   REGISTER = "Register",
   LOG_IN = "Log in",
   CREATE_EXPENSE = "Create expense",
   GET_ALL_EXPENSES = "Get all my expenses",
   GET_EXPENSE = "Get expense by ID",
   UPDATE_EXPENSE = "Update expense by ID",
   DELETE_EXPENSE = "Delete expense by ID",
   DELETE_ACCOUNT = "Delete my account",
   REFRESH_TOKEN = "Refresh token"
}